requirements.txt
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dialogflow==0.6.0
enum34==1.1.6
feedparser==5.2.1
Flask==1.0.3
futures==3.1.1
fuzzywuzzy==0.17.0
gnewsclient==1.12
google-api-core==1.11.1
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.21.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyJWT==1.7.1
PySocks==1.7.0
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
twilio==6.27.1
urllib3==1.25.3
Werkzeug==0.15.4
